requirements.txt
pymongo==3.12.1
streamlit
Pillow==8.4.0
dnspython